fix: dissociations in Car, Model and RentalRate entity session beans

+++ CarRentalManagementSystem-ejb/src/java/ejb/session/stateless/ModelEntitySessionBean.java
@@ -1,8 +1,11 @@
 package ejb.session.stateless;
 
+import entity.CarEntity;
 import entity.CategoryEntity;
 import entity.ModelEntity;
 import java.util.List;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import javax.ejb.EJB;
 import javax.ejb.Local;
 import javax.ejb.Remote;
@@ -10,6 +13,8 @@ import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
+import util.enumeration.StatusEnum;
+import util.exception.CarNotFoundException;
 import util.exception.CategoryNotFoundException;
 import util.exception.ModelNotFoundException;
 
@@ -24,6 +29,9 @@ public class ModelEntitySessionBean implements ModelEntitySessionBeanRemote, Mod
 
     @EJB
     private CategoryEntitySessionBeanLocal categoryEntitySessionBeanLocal;
+
+    @EJB
+    private CarEntitySessionBeanLocal carEntitySessionBeanLocal;
 
     @Override
     public ModelEntity createNewModelEntity(ModelEntity newModelEntity, Long categoryId) throws CategoryNotFoundException {
@@ -57,18 +65,33 @@ public class ModelEntitySessionBean implements ModelEntitySessionBeanRemote, Mod
             throw new ModelNotFoundException("Model ID " + modelId + " does not exist!");
         }
     }
-    
+
     @Override
     public void updateModel(ModelEntity modelEntity) {
         em.merge(modelEntity);
         em.flush();
     }
-    
+
     @Override
     public void deleteModel(Long modelId) throws ModelNotFoundException {
         ModelEntity modelEntityToRemove = retrieveModelEntityByModelId(modelId);
-        em.remove(modelEntityToRemove);
+        modelEntityToRemove.getCategoryEntity().getModelEntities().remove(modelEntityToRemove);
+
+        // delete all cars of this model (or set disabled, if alr used)
+        for (CarEntity carEntity : modelEntityToRemove.getCarEntities()) {
+            try {
+                carEntitySessionBeanLocal.deleteCar(carEntity.getCarId());
+            } catch (CarNotFoundException ex) {
+                System.out.println("Error deleting car ID " + carEntity.getCarId() + ": " + ex.getMessage());
+            }
+        }
+        
+        if (modelEntityToRemove.getModelStatus() == StatusEnum.USED) {
+            modelEntityToRemove.setModelStatus(StatusEnum.DISABLED);
+        } else {
+            em.remove(modelEntityToRemove);
+        }
+        
         em.flush();
     }
-
 }
